backend/QLunch/requirements.txt
asgiref==3.9.1
boto3==1.40.5
botocore==1.40.5
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
django-rest-passwordreset==1.5.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.9.0
